refactored TIFDataService and TIFIndexService classes
