+++ src/main/java/org/opensearch/securityanalytics/threatIntel/ThreatIntelFeedDataService.java
@@ -4,6 +4,7 @@
  */
 package org.opensearch.securityanalytics.threatIntel;
 
+import org.apache.commons.csv.CSVParser;
 import org.apache.commons.csv.CSVRecord;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.LogManager;
@@ -35,9 +36,12 @@ import org.opensearch.core.xcontent.ToXContent;
 import org.opensearch.core.xcontent.XContentBuilder;
 import org.opensearch.securityanalytics.model.ThreatIntelFeedData;
 import org.opensearch.securityanalytics.threatIntel.action.ThreatIntelIndicesResponse;
+import org.opensearch.securityanalytics.threatIntel.common.TIFJobState;
 import org.opensearch.securityanalytics.threatIntel.common.TIFMetadata;
 import org.opensearch.securityanalytics.threatIntel.common.StashedThreadContext;
 import org.opensearch.securityanalytics.settings.SecurityAnalyticsSettings;
+import org.opensearch.securityanalytics.threatIntel.feedMetadata.BuiltInTIFMetadataLoader;
+import org.opensearch.securityanalytics.threatIntel.jobscheduler.TIFJobSchedulerMetadata;
 import org.opensearch.securityanalytics.threatIntel.jobscheduler.TIFJobSchedulerMetadataService;
 import org.opensearch.securityanalytics.util.IndexUtils;
 import org.opensearch.securityanalytics.util.SecurityAnalyticsException;
@@ -47,13 +51,9 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.nio.charset.StandardCharsets;
+import java.time.Duration;
 import java.time.Instant;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
+import java.util.*;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
@@ -64,34 +64,29 @@ import static org.opensearch.securityanalytics.threatIntel.jobscheduler.TIFJobSc
  * Service to handle CRUD operations on Threat Intel Feed Data
  */
 public class ThreatIntelFeedDataService {
-
     private static final Logger log = LogManager.getLogger(ThreatIntelFeedDataService.class);
-
-    public static final String SETTING_INDEX_REFRESH_INTERVAL = "index.refresh_interval";
-    private static final Map<String, Object> INDEX_SETTING_TO_CREATE = Map.of(
-            IndexMetadata.SETTING_NUMBER_OF_SHARDS,
-            1,
-            IndexMetadata.SETTING_NUMBER_OF_REPLICAS,
-            0,
-            SETTING_INDEX_REFRESH_INTERVAL,
-            -1,
-            IndexMetadata.SETTING_INDEX_HIDDEN,
-            true
-    );
-
+    private static final int SLEEP_TIME_IN_MILLIS = 5000; // 5 seconds
+    private static final int MAX_WAIT_TIME_FOR_REPLICATION_TO_COMPLETE_IN_MILLIS = 10 * 60 * 60 * 1000; // 10 hours
     private final ClusterService clusterService;
     private final ClusterSettings clusterSettings;
     private final NamedXContentRegistry xContentRegistry;
     private final Client client;
+    private final BuiltInTIFMetadataLoader builtInTIFMetadataLoader;
+    private final TIFJobSchedulerMetadataService tifJobSchedulerMetadataService;
+
 
     public ThreatIntelFeedDataService(
             ClusterService clusterService,
             Client client,
-            NamedXContentRegistry xContentRegistry) {
+            NamedXContentRegistry xContentRegistry,
+            BuiltInTIFMetadataLoader builtInTIFMetadataLoader,
+            TIFJobSchedulerMetadataService tifJobSchedulerMetadataService) {
         this.client = client;
         this.xContentRegistry = xContentRegistry;
         this.clusterService = clusterService;
         this.clusterSettings = clusterService.getClusterSettings();
+        this.builtInTIFMetadataLoader = builtInTIFMetadataLoader;
+        this.tifJobSchedulerMetadataService = tifJobSchedulerMetadataService;
     }
 
     public void getThreatIntelFeedData(
@@ -107,31 +102,94 @@ public class ThreatIntelFeedDataService {
         }));
     }
 
-    /**
-     * Create an index for a threat intel feed
-     * <p>
-     * Index setting start with single shard, zero replica, no refresh interval, and hidden.
-     * Once the threat intel feed is indexed, do refresh and force merge.
-     * Then, change the index setting to expand replica to all nodes, and read only allow delete.
-     *
-     * @param indexName index name
-     */
-    public void createIndexIfNotExists(final String indexName, final ActionListener<CreateIndexResponse> listener) {
-        if (clusterService.state().metadata().hasIndex(indexName) == true) {
-            listener.onResponse(new CreateIndexResponse(true, true, indexName));
-            return;
+    public GroupedActionListener<CreateIndexResponse> getCreateIndexResponseGroupedActionListener(
+            TIFJobSchedulerMetadata tifJobSchedulerMetadata,
+            Runnable renewLock, ActionListener<ThreatIntelIndicesResponse> listener,
+            List<AbstractMap.SimpleEntry<TIFJobSchedulerMetadata, TIFMetadata>> tifMetadataList
+    ) {
+        Instant startTime = Instant.now();
+        Map<String, TIFMetadata> indexTIFMetadataMap = new HashMap<>();
+        for (TIFMetadata tifMetadata: builtInTIFMetadataLoader.getTifMetadataList()) {
+            String indexName = tifJobSchedulerMetadata.newIndexName(tifJobSchedulerMetadata, tifMetadata);
+            tifMetadataList.add(new AbstractMap.SimpleEntry<>(tifJobSchedulerMetadata, tifMetadata));
+            indexTIFMetadataMap.put(indexName, tifMetadata);
         }
-        final CreateIndexRequest createIndexRequest = new CreateIndexRequest(indexName).settings(INDEX_SETTING_TO_CREATE)
-                .mapping(getIndexMapping()).timeout(clusterSettings.get(SecurityAnalyticsSettings.THREAT_INTEL_TIMEOUT));
-        StashedThreadContext.run(
-                client,
-                () -> client.admin().indices().create(createIndexRequest, new ActionListener<>() {
+
+        GroupedActionListener<CreateIndexResponse> createdThreatIntelIndices = new GroupedActionListener<>(
+                new ActionListener<>() {
                     @Override
-                    public void onResponse(CreateIndexResponse response) {
-                        if (response.isAcknowledged()) {
-                            listener.onResponse(response);
-                        } else {
-                            onFailure(new OpenSearchStatusException("Threat intel feed index creation failed", RestStatus.INTERNAL_SERVER_ERROR));
+                    public void onResponse(Collection<CreateIndexResponse> responses) {
+                        try {
+
+                            int noOfUnprocessedResponses = 0;
+                            for (CreateIndexResponse response: responses) {
+                                String indexName = response.index();
+                                TIFMetadata tifMetadata = indexTIFMetadataMap.get(indexName);
+                                if (tifMetadata.getFeedType().equals("csv")) {
+                                    ++noOfUnprocessedResponses;
+                                }
+                            }
+                            GroupedActionListener<ThreatIntelIndicesResponse> saveThreatIntelFeedResponseListener = new GroupedActionListener<>(new ActionListener<>() {
+                                @Override
+                                public void onResponse(Collection<ThreatIntelIndicesResponse> responses) {
+                                    List<String> freshIndices = new ArrayList<>();
+                                    for (ThreatIntelIndicesResponse response: responses) {
+                                        Boolean succeeded = false;
+                                        if (response.isAcknowledged()) {
+                                            String indexName = response.getIndices().get(0);
+                                            waitUntilAllShardsStarted(indexName, MAX_WAIT_TIME_FOR_REPLICATION_TO_COMPLETE_IN_MILLIS);
+                                            freshIndices.add(indexName);
+                                            succeeded = true;
+                                        }
+
+                                        if (!succeeded) {
+                                            log.error("Exception: failed to parse correct feed type");
+                                            onFailure(new OpenSearchException("Exception: failed to parse correct feed type"));
+                                        }
+                                    }
+
+                                    Instant endTime = Instant.now();
+                                    updateJobSchedulerMetadataAsSucceeded(freshIndices, tifJobSchedulerMetadata, startTime, endTime, listener);
+                                }
+
+                                @Override
+                                public void onFailure(Exception e) {
+                                    listener.onFailure(e);
+                                }
+                            }, noOfUnprocessedResponses);
+
+                            for (CreateIndexResponse response: responses) {
+                                String indexName = response.index();
+                                TIFMetadata tifMetadata = indexTIFMetadataMap.get(indexName);
+                                switch (tifMetadata.getFeedType()) {
+                                    case "csv":
+                                        try (CSVParser reader = ThreatIntelFeedParser.getThreatIntelFeedReaderCSV(tifMetadata)) {
+                                            CSVParser noHeaderReader = ThreatIntelFeedParser.getThreatIntelFeedReaderCSV(tifMetadata);
+                                            boolean notFound = true;
+
+                                            while (notFound) {
+                                                CSVRecord hasHeaderRecord = reader.iterator().next();
+
+                                                //if we want to skip this line and keep iterating
+                                                if ((hasHeaderRecord.values().length ==1 && "".equals(hasHeaderRecord.values()[0])) || hasHeaderRecord.get(0).charAt(0) == '#' || hasHeaderRecord.get(0).charAt(0) == ' '){
+                                                    noHeaderReader.iterator().next();
+                                                } else { // we found the first line that contains information
+                                                    notFound = false;
+                                                }
+                                            }
+                                            if (tifMetadata.hasHeader()){
+                                                parseAndSaveThreatIntelFeedDataCSV(indexName, reader.iterator(), renewLock, tifMetadata, saveThreatIntelFeedResponseListener);
+                                            } else {
+                                                parseAndSaveThreatIntelFeedDataCSV(indexName, noHeaderReader.iterator(), renewLock, tifMetadata, saveThreatIntelFeedResponseListener);
+                                            }
+                                        }
+                                        break;
+                                    default:
+                                        // if the feed type doesn't match any of the supporting feed types, throw an exception
+                                }
+                            }
+                        } catch (IOException ex) {
+                            onFailure(ex);
                         }
                     }
 
@@ -139,8 +197,10 @@ public class ThreatIntelFeedDataService {
                     public void onFailure(Exception e) {
                         listener.onFailure(e);
                     }
-                })
+                },
+                tifMetadataList.size()
         );
+        return createdThreatIntelIndices;
     }
 
     /**
@@ -243,55 +303,55 @@ public class ThreatIntelFeedDataService {
 
     }
 
-    public void deleteThreatIntelDataIndex(final List<String> indices) {
-        if (indices == null || indices.isEmpty()) {
-            return;
-        }
-
-        Optional<String> invalidIndex = indices.stream()
-                .filter(index -> index.startsWith(THREAT_INTEL_DATA_INDEX_NAME_PREFIX) == false)
-                .findAny();
-        if (invalidIndex.isPresent()) {
+    /**
+     * We wait until all shards are ready to serve search requests before updating job scheduler parameter to
+     * point to a new index so that there won't be latency degradation during threat intel feed data update
+     *
+     * @param indexName the indexName
+     */
+    protected void waitUntilAllShardsStarted(final String indexName, final int timeout) {
+        Instant start = Instant.now();
+        try {
+            while (Instant.now().toEpochMilli() - start.toEpochMilli() < timeout) {
+                if (clusterService.state().routingTable().allShards(indexName).stream().allMatch(shard -> shard.started())) {
+                    return;
+                }
+                Thread.sleep(SLEEP_TIME_IN_MILLIS);
+            }
             throw new OpenSearchException(
-                    "the index[{}] is not threat intel data index which should start with {}",
-                    invalidIndex.get(),
-                    THREAT_INTEL_DATA_INDEX_NAME_PREFIX
+                    "index[{}] replication did not complete after {} millis",
+                    MAX_WAIT_TIME_FOR_REPLICATION_TO_COMPLETE_IN_MILLIS
             );
+        } catch (InterruptedException e) {
+            log.error("runtime exception", e);
+            throw new SecurityAnalyticsException("Runtime exception", RestStatus.INTERNAL_SERVER_ERROR, e);
         }
-
-        StashedThreadContext.run(
-                client,
-                () -> client.admin()
-                        .indices()
-                        .prepareDelete(indices.toArray(new String[0]))
-                        .setIndicesOptions(IndicesOptions.LENIENT_EXPAND_OPEN_CLOSED_HIDDEN)
-                        .setTimeout(clusterSettings.get(SecurityAnalyticsSettings.THREAT_INTEL_TIMEOUT))
-                        .execute(new ActionListener<>() {
-                            @Override
-                            public void onResponse(AcknowledgedResponse response) {
-                                if (response.isAcknowledged() == false) {
-                                    onFailure(new OpenSearchException("failed to delete data[{}]", String.join(",", indices)));
-                                }
-                            }
-
-                            @Override
-                            public void onFailure(Exception e) {
-                                log.error("unknown exception:", e);
-                            }
-                        })
-        );
     }
 
-    private String getIndexMapping() {
-        try {
-            try (InputStream is = TIFJobSchedulerMetadataService.class.getResourceAsStream("/mappings/threat_intel_feed_mapping.json")) {
-                try (BufferedReader reader = new BufferedReader(new InputStreamReader(is, StandardCharsets.UTF_8))) {
-                    return reader.lines().map(String::trim).collect(Collectors.joining());
-                }
-            }
-        } catch (IOException e) {
-            log.error("Runtime exception when getting the threat intel index mapping", e);
-            throw new SecurityAnalyticsException("Runtime exception when getting the threat intel index mapping", RestStatus.INTERNAL_SERVER_ERROR, e);
-        }
+    /**
+     * Update tifJobSchedulerMetadata as succeeded
+     * @param tifJobSchedulerMetadata the tifJobSchedulerMetadata
+     * @param startTime the start time
+     * @param endTime the end time
+     * @param listener the action listener
+     */
+    public void updateJobSchedulerMetadataAsSucceeded(
+            List<String> indices,
+            final TIFJobSchedulerMetadata tifJobSchedulerMetadata,
+            final Instant startTime,
+            final Instant endTime,
+            final ActionListener<ThreatIntelIndicesResponse> listener
+    ) {
+        tifJobSchedulerMetadata.setIndices(indices);
+        tifJobSchedulerMetadata.getUpdateStats().setLastSucceededAt(endTime);
+        tifJobSchedulerMetadata.getUpdateStats().setLastProcessingTimeInMillis(endTime.toEpochMilli() - startTime.toEpochMilli());
+        tifJobSchedulerMetadata.enable();
+        tifJobSchedulerMetadata.setState(TIFJobState.AVAILABLE);
+        tifJobSchedulerMetadataService.updateJobSchedulerMetadata(tifJobSchedulerMetadata, listener);
+        log.info(
+                "threat intel feed data creation succeeded for {} and took {} seconds",
+                tifJobSchedulerMetadata.getName(),
+                Duration.between(startTime, endTime)
+        );
     }
 }
